Added message to asserts, moved UtilTest to utils package, removed static login details from UtilTest, UtilTest uses testdata.properties, refactored methods to better names
- readProductName -> getProductName
- getShoppingCartAmount -> getProductsCount
- login -> loginAsStandardUser
