+++ src/main/java/com/mycompany/carina/demo/gui/saucedemo/component/ProductItem.java
@@ -28,7 +28,7 @@ public class ProductItem extends AbstractUIObject {
         addToCartButton.click();
     }
 
-    public String readProductName() {
+    public String getProductName() {
         return itemName.getText();
     }
 }
