+++ src/main/java/com/mycompany/carina/demo/gui/saucedemo/pages/ProductPage.java
@@ -63,7 +63,7 @@ public class ProductPage extends AbstractPage {
     }
 
     public ProductItem findProductByName(String name) {
-        return products.stream().filter(e -> e.readProductName().equalsIgnoreCase(name)).findFirst().get();
+        return products.stream().filter(e -> e.getProductName().equalsIgnoreCase(name)).findFirst().get();
     }
 
     public int getShoppingCartAmount() {
