+++ src/main/java/com/mycompany/carina/demo/gui/saucedemo/pages/checkout/CheckoutOverviewPage.java
@@ -1,5 +1,6 @@
 package com.mycompany.carina.demo.gui.saucedemo.pages.checkout;
 
+import com.mycompany.carina.demo.gui.saucedemo.component.ProductItem;
 import com.qaprosoft.carina.core.foundation.webdriver.decorator.ExtendedWebElement;
 import com.qaprosoft.carina.core.gui.AbstractPage;
 import org.openqa.selenium.WebDriver;
@@ -8,6 +9,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.lang.invoke.MethodHandles;
+import java.util.List;
 
 public class CheckoutOverviewPage extends AbstractPage {
     private static final Logger LOGGER = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
@@ -16,6 +18,9 @@ public class CheckoutOverviewPage extends AbstractPage {
     @FindBy(xpath = "//*[@id='finish']")
     private ExtendedWebElement finishButton;
 
+    @FindBy(className = "cart_item")
+    private List<ProductItem> products;
+
     public CheckoutOverviewPage(WebDriver driver) {
         super(driver);
         setPageAbsoluteURL(pageUrl);
@@ -23,5 +28,9 @@ public class CheckoutOverviewPage extends AbstractPage {
 
     public void clickFinishButton() {
         finishButton.click();
+    }
+
+    public int getProductsCount() {
+        return products.size();
     }
 }
