+++ src/main/resources/_testdata.properties
@@ -2,3 +2,5 @@
 #========================== Test data ==========================#
 #===============================================================#
 test_credentials=test@gmail.com/{crypt:8O9iA4+f3nMzz85szmvKmQ==}
+test_username=standard_user
+test_password=secret_sauce
