+++ src/test/java/com/mycompany/carina/demo/mytests/AbstractSauceDemoTest.java
@@ -1,5 +1,6 @@
 package com.mycompany.carina.demo.mytests;
 
+import com.mycompany.carina.demo.utils.UtilTest;
 import com.qaprosoft.carina.core.foundation.IAbstractTest;
 
 public abstract class AbstractSauceDemoTest implements IAbstractTest {
