+++ src/test/java/com/mycompany/carina/demo/mytests/SauceDemoTest.java
@@ -5,7 +5,6 @@ import com.mycompany.carina.demo.gui.saucedemo.pages.CartPage;
 import com.mycompany.carina.demo.gui.saucedemo.pages.CheckoutYourInformationPage;
 import com.mycompany.carina.demo.gui.saucedemo.pages.ProductPage;
 import com.mycompany.carina.demo.gui.saucedemo.pages.SauceLabsPage;
-import com.mycompany.carina.demo.gui.saucedemo.pages.checkout.CheckoutCompletePage;
 import com.mycompany.carina.demo.gui.saucedemo.pages.checkout.CheckoutOverviewPage;
 import org.testng.Assert;
 import org.testng.annotations.Test;
@@ -13,14 +12,14 @@ import org.testng.annotations.Test;
 public class SauceDemoTest extends AbstractSauceDemoTest {
     @Test
     public void testSort() {
-        ProductPage productPage = utilTest.login();
+        ProductPage productPage = utilTest.loginAsStandardUser();
         productPage.setSortOption(SortOptions.LOW_TO_HIGH);
-        Assert.assertEquals(productPage.getSortActiveOption(), SortOptions.LOW_TO_HIGH);
+        Assert.assertEquals(productPage.getSortActiveOption(), SortOptions.LOW_TO_HIGH, "Sort Bar active option is not low to high.");
     }
 
     @Test
     public void testAboutPage() {
-        ProductPage productPage = utilTest.login();
+        ProductPage productPage = utilTest.loginAsStandardUser();
         productPage.navigate("About");
         SauceLabsPage sauceLabsPage = new SauceLabsPage((getDriver()));
         Assert.assertTrue(sauceLabsPage.isPageOpened(), "Sauce labs page is not opened!");
@@ -29,28 +28,29 @@ public class SauceDemoTest extends AbstractSauceDemoTest {
     @Test
     public void testAddCart() {
         String product = "Sauce Labs Backpack";
-        ProductPage productPage = utilTest.login();
+        ProductPage productPage = utilTest.loginAsStandardUser();
         productPage.clickProductAddToCart(product);
         productPage.clickShoppingCart();
         CartPage cartPage = new CartPage(getDriver());
-        Assert.assertEquals(cartPage.findProductByName(product).readProductName(), product);
+        Assert.assertEquals(cartPage.findProductByName(product).getProductName(), product, "Unable to find product in cart page.");
     }
 
     @Test
     public void testRemoveCart() {
         String product = "Sauce Labs Backpack";
-        ProductPage productPage = utilTest.login();
+        ProductPage productPage = utilTest.loginAsStandardUser();
         productPage.clickProductAddToCart(product);
         productPage.clickShoppingCart();
         CartPage cartPage = new CartPage(getDriver());
         cartPage.clickProductRemoveFromCart(product);
-        Assert.assertEquals(cartPage.getShoppingCartAmount(), 0);
+        Assert.assertEquals(cartPage.getProductsCount(), 0, "Unable to remove product in cart page.");
     }
 
     @Test
     public void testCheckout() {
         String product = "Sauce Labs Backpack";
-        ProductPage productPage = utilTest.login();
+        ProductPage productPage = utilTest.loginAsStandardUser();
+        productPage.clickProductAddToCart(product);
         productPage.clickShoppingCart();
         CartPage cartPage = new CartPage(getDriver());
         cartPage.clickCheckOutButton();
@@ -60,9 +60,8 @@ public class SauceDemoTest extends AbstractSauceDemoTest {
         checkoutYourInformationPage.typePostalCodeInput("12345");
         checkoutYourInformationPage.clickContinueButton();
         CheckoutOverviewPage checkoutOverviewPage = new CheckoutOverviewPage(getDriver());
-        checkoutOverviewPage.clickFinishButton();
-        CheckoutCompletePage checkoutCompletePage = new CheckoutCompletePage(getDriver());
-        Assert.assertTrue(checkoutCompletePage.isPageOpened(), "Checkout is not complete!");
+        // assert on this page
+        Assert.assertEquals(checkoutOverviewPage.getProductsCount(), 1, "Incorrect amount of products found.");
     }
 
 }
